requirements.txt
Kivy==2.1.0
Kivy-Garden==0.1.5
KivyAuth==2.3.3
kivymd==1.1.1
oauthlib==3.2.2
requests==2.28.2
mapview==1.0.6
python-dotenv==1.0.0
cryptography==39.0.2
keyring==23.13.1
multitasking==0.0.11
keyring==23.13.1